✨ Filtro no front pelo nome do cliente.
